% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{coef_store}
\alias{coef_store}
\alias{cl_store}
\title{Store for fit-coefficients}
\usage{
cl_store()
}
\description{
For an interactive analysis of fits using \code{\link{mbte_visualize_coef}},
filtering based on model coefficients may be required. The mentioned options
below are available for storing model coefficients.
}
\details{
A coefficient store should provide the following functions (accessible via
the \code{$}-operator):
\describe{
  \item{add_row(...)}{Add model coefficients (ellipsis must be named).}
  \item{get_tibble()}{Return a \code{\link[tibble]{tibble}} with the added
    model coefficients (the order, in which the coefficients were added, is
    preserved).}
}
}
\section{Functions}{
\itemize{
\item \code{cl_store}: Store fitting-coefficients using a closure (rely on
lexical scoping).
}}

\examples{
# create closure-based coefficient-store
store <- cl_store()

# store coefficients
store$add_row(A = 10, B = -5, C = 0)
store$add_row(A = -1, B = 3, C = 1)

# retrieve tibble containing coefficients
store$get_tibble()

}
