% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mbte_coef_fit}
\alias{mbte_coef_fit}
\title{A wrapper to include fit coefficients as a separate column}
\usage{
mbte_coef_fit(x, ...)
}
\arguments{
\item{x}{A \code{\link{tbl_mbte}}.}

\item{...}{Trend modules to use.}
}
\value{
A \code{\link[mbte:tbl_mbte]{tbl_mbte}} with fit coefiicients added.
}
\description{
\code{\link[mbte:mbte_fit]{mbte_fit}} is wrapped. Currently, the column
\code{coefficients} is added to the resulting tibble.
}
\note{
It is assumed, that each trend module adds coefficients to the
  corresponding coefficient store exactly once. If fitting fails, NA's
  should be added.
}
