% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{mbte_visualize_coef}
\alias{mbte_visualize_coef}
\title{Visualizae coefficient-based fits using trelliscopejs package}
\usage{
mbte_visualize_coef(x, ...)
}
\arguments{
\item{x}{A \code{\link{tbl_mbte}}.}

\item{...}{Columns to use for facetting (see
\code{\link[trelliscopejs]{facet_trelliscope}}).}
}
\description{
Visualizae coefficient-based fits using trelliscopejs package
}
