% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-modules.R, R/linear.R, R/logistic.R
\name{trend-modules}
\alias{trend-modules}
\alias{tm_linear}
\alias{tm_logistic}
\title{Trend modules for coefficient based fitting}
\usage{
tm_linear(coef_store = cl_store())

tm_logistic(coef_store = cl_store())
}
\arguments{
\item{coef_store}{A \code{\link[mbte.vis:coef_store]{coefficient-store}} to
record fit parameters.}
}
\description{
Trend modules are wrappers around
\code{\link[mbte:fitting-helpers]{trend fitting helpers}} in the mbte-
package. They are meant to be used with \code{\link{mbte_coef_fit}}.
}
\details{
A fitting module consists of the following components:
\describe{
  \item{fitting quosure}{Adapted fitting quosure, based on the correspoding
    fitting helper in the mbte-package.}
  \item{coefficient store}{Stores fit coefficients.}
}
}
\section{Functions}{
\itemize{
\item \code{tm_linear}: Fit a linear trend using
\code{\link[mbte]{tr_linear}}. Fit coefficients:
\itemize{
  \item prefix_slope
  \item prefix_rel_slope (relative slope - normed to signal maximum)
  \item prefix_intercept
  \item prefix_rel_intercept
}

\item \code{tm_logistic}: Fit a logistic trend using
\code{\link[mbte]{tr_logistic}}. Equation used for fitting:
\code{value ~ A / (1 + exp(B * (C - time))) + D}. Fit coefficients:
\itemize{
  \item prefix_A
  \item prefix_rel_A (relative A - normed to signal maximum)
  \item prefix_B
  \item prefix_rel_B
  \item prefix_C
  \item prefix_rel_C
  \item prefix_D
  \item prefix_rel_D
}
}}

\examples{
# use dataset from the mbte package
library(mbte)
data(filtered_signals)

# perform fitting (record the coefficients)
mbte_coef_fit(filtered_signals, lin = tm_linear(), log = tm_logistic())

# equivalent using the mbte-package (don't store coefficients)
mbte_fit(filtered_signals, lin = !!tr_linear(), log = !!tr_logistic)

}
